Module rs_poker::core

Allow all the core poker functionality to be used externally. Everything in core should be agnostic to poker style. This is the core module. It exports the non-holdem related code.

Structs

Card
The main struct of this library. This is a carrier for Suit and Value combined.

CardIter
Given some cards create sets of possible groups of cards.

Deck
Deck struct that can tell quickly if a card is in the deck

FlatDeck
FlatDeck is a deck of cards that allows easy indexing into the cards. It does not provide contains methods.

Hand
Struct to hold cards.

Enums

Rank
All the different possible hand ranks. For each hand rank the u32 corresponds to the strength of the hand in comparison to others of the same rank.

Suit
Enum for the four different suits. While this has support for ordering it's not sensical. The sorting is only there to allow sorting cards.

Value
Card rank or value. This is basically the face value - 2

Traits

Flattenable
Trait that means a deck can be made into a FlatDeck

Rankable
Can this turn into a hand rank? There are default implementations for Hand and Vec<Card>.