Struct rs_poker::core::FlatDeck

pub struct FlatDeck { /* fields omitted */ }

FlatDeck is a deck of cards that allows easy indexing into the cards. It does not provide contains methods.

Methods

impl FlatDeck

pub fn len(&self) -> usize

How many cards are there in the deck ?

pub fn is_empty(&self) -> bool

Have all cards been dealt ? This probably won't be used as it's unlikely that someone will deal all 52 cards from a deck.

pub fn sample(&self, n: usize) -> Vec<Card>

Give a random sample of the cards still left in the deck

pub fn shuffle(&mut self)

Randomly shuffle the flat deck. This will ensure the there's no order to the deck.

pub fn deal(&mut self) -> Option<Card>

Deal a card if there is one there to deal. None if the deck is empty