Struct rs_poker::core::Hand

pub struct Hand { /* fields omitted */ }

Struct to hold cards.

This doesn't have the ability to easily check if a card is in the hand. So do that before adding/removing a card.

Methods

impl Hand

pub fn default() -> Hand

Create the default empty hand.

pub fn new_with_cards(cards: Vec<Card>) -> Hand

Create the hand with specific hand.

pub fn new_from_str(hand_string: &str) -> Result<Hand, String>

From a str create a new hand.

Examples

use rs_poker::core::Hand;
let hand = Hand::new_from_str("AdKd").unwrap();

Anything that can't be parsed will return an error.

use rs_poker::core::Hand;
let hand = Hand::new_from_str("AdKx");
assert!(hand.is_err());

pub fn push(&mut self, c: Card)

Add card at to the hand. No verification is done at all.

pub fn truncate(&mut self, len: usize)

Truncate the hand to the given number of cards.

pub fn len(&self) -> usize

How many cards are in this hand so far ?

pub fn is_empty(&self) -> bool

Are there any cards at all ?

pub fn iter(&self) -> Iter<Card>

Create an iter on the cards.