Module rs_poker::holdem

The holdem specific code. This contains range parsing, game state, and starting hand code.

Structs

MonteCarloGame
Current state of a game.

RangeParser
Unit struct to provide starting hand parse functions. Use this to parse things like RangeParser::parse_one("AKo") and RangeParser::parse_one("TT+")

Enums

StartingHand
Enum to represent all the possible ways to specify a starting hand.

Suitedness
Enum to represent how the suits of a hand correspond to each other. Suitedness::Suited will mean that all cards have the same suit Suitedness::OffSuit will mean that all cards have the different suit Suitedness::Any makes no promises.