Enum rs_poker::holdem::StartingHand

pub enum StartingHand {
    Def(Default),
    SingleCardRange(SingleCardRange),
}

Enum to represent all the possible ways to specify a starting hand.

Variants
Def(Default)

Default starting hand type. This means that we specify two cards and their suitedness.

SingleCardRange(SingleCardRange)

A starting hand where the second card is a range.

Methods

impl StartingHand

pub fn default(
value_one: Value,
value_two: Value,
suited: Suitedness
) -> StartingHand

Create a default starting hand with two Value's and a Suitedness.

pub fn single_range(
value_one: Value,
start: Value,
end: Value,
suited: Suitedness
) -> StartingHand

Create a new StartingHand with the second card being a range.

pub fn all() -> Vec<StartingHand>

Create every possible unique StartingHand.

pub fn possible_hands(&self) -> Vec<Hand>

From a StartingHand specify all the hands this could represent.