Trait rs_poker::core::Rankable

pub trait Rankable {
    fn cards(&self) -> &[Card];

    fn rank(&self) -> Rank { ... }

    fn rank_five(&self) -> Rank { ... }
}

Can this turn into a hand rank? There are default implementations for Hand and Vec<Card>.

Required Methods

fn cards(&self) -> &[Card]

Rank the current 5 card hand. This will no cache the value.

Provided Methods

fn rank(&self) -> Rank

Rank the cards to find the best 5 card hand. This will work on 5 cards or more ( specifically on 7 card holdem hands). If you know that the hand only contains 5 cards then rank_five will be faster.

Examples

use rs_poker::core::{Hand, Rank, Rankable};

let hand = Hand::new_from_str("2h2d8d8sKd6sTh").unwrap();
let rank = hand.rank();
assert!(Rank::TwoPair(0) <= rank);
assert!(Rank::TwoPair(u32::max_value()) >= rank);

fn rank_five(&self) -> Rank

Rank this hand. It doesn't do any caching so it's left up to the user to understand that duplicate work will be done if this is called more than once.