Struct rs_poker::holdem::RangeParser

pub struct RangeParser;

Unit struct to provide starting hand parse functions. Use this to parse things like RangeParser::parse_one("AKo") and RangeParser::parse_one("TT+")

Methods

impl RangeParser

pub fn parse_one(r_str: &str) -> Result<Vec<Hand>, String>

Parse a string and return all the starting hands

Examples

You can send in hands where the suits are specified for all hands.

use rs_poker::holdem::RangeParser;

let hand = RangeParser::parse_one("AhKh").unwrap();
assert_eq!(1, hand.len());

You can also specify hands were the suits are not specified, but you want them to be suited.

use rs_poker::holdem::RangeParser;
use rs_poker::core::Value;

let hands = RangeParser::parse_one("AKs").unwrap();
assert!(hands.len() == 4);
for hand in hands {
    assert!(hand[0].suit == hand[1].suit);
    assert_eq!(Value::Ace, hand[0].value);
    assert_eq!(Value::King, hand[1].value);
}

You can also specify that the cards are not of the same suit.

use rs_poker::holdem::RangeParser;
use rs_poker::core::Value;

let hands = RangeParser::parse_one("AKo").unwrap();

for hand in hands {
    assert!(hand[0].suit != hand[1].suit);
    assert_eq!(Value::Ace, hand[0].value);
    assert_eq!(Value::King, hand[1].value);
}

You can also use the + modifier after a set of starting cards. The modifier will mean different things for different sets of cards.

If the starting cards are pairs then the + means all pairs equal to or above the specified values.

use rs_poker::holdem::RangeParser;
use rs_poker::core::Value;

let hands = RangeParser::parse_one("TT+").unwrap();

for hand in hands {
    assert!(hand[0].value == hand[1].value);
    assert!(hand[0].value >= Value::Ten);
    assert!(hand[1].value >= Value::Ten);
}

If the cards are connectors then the plus means all connectors where the cards are above the specified values.

use rs_poker::holdem::RangeParser;
use rs_poker::core::Value;

let hands = RangeParser::parse_one("T9o+").unwrap();

for hand in hands {
    assert_eq!(hand[0].value, Value::from_u8(hand[1].value as u8 + 1));
    assert!(hand[0].suit != hand[1].suit);
}

If the cards are not paired and not connectors then plus after the hand means all hands where the second card is greater than or equal to the specified second card, and below the first card.

use rs_poker::holdem::RangeParser;
use rs_poker::core::Value;
let hands = RangeParser::parse_one("A9s+").unwrap();
for hand in hands {
    assert!(hand[0].value > hand[1].value);
    assert!(hand[1].value >= Value::Nine);
    assert!(hand[0].suit == hand[1].suit);
}

It's also possible to do more complex ranges using dash modifer. For example if you wanted to represent Suited middle connectors you could do something like this.

use rs_poker::holdem::RangeParser;
use rs_poker::core::Value;

let hands = RangeParser::parse_one("JT-67s").unwrap();
for hand in hands {
    // The largest card is always the first
    assert!(hand[0].value > hand[1].value);
    // first card is great or equal to seven
    assert!(hand[0].value >= Value::Seven);
    // Second card is greater or equal to six
    assert!(hand[1].value >= Value::Six);
    // First card is less than or equal to Jack
    assert!(hand[0].value <= Value::Jack);
    // Second card is less than or equal to Ten
    assert!(hand[1].value <= Value::Ten);
    // All hands are connectors.
    assert_eq!(1, hand[0].value.gap(&hand[1].value));
}

Also with the dash modifier there's no need to only have connected cards. It's possible to represent ranges with gappers. For example if you wanted to do high one gappers.

use rs_poker::holdem::RangeParser;
let hands = RangeParser::parse_one("AQ-J9").unwrap();
println!("Hands = {:?}", hands);

Since the dash modifier represents a range the difference between cards ( the gap ) must remain constant. If it's not parse_one will return an Err.

use rs_poker::holdem::RangeParser;
// This will not work since the difference between Ace and King is one
// while the diffence between Jack and Nine is two.
let hands = RangeParser::parse_one("AK-J9").unwrap();
// We'll never get here
println!("Hands = {:?}", hands);